improve verifyHeader
